+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Introduced ``Repository.is_temporary`` attribute
+- Support get maintenance status of pulp repositories
 
 ## [1.2.1] - 2019-08-12